# src/main/resources/doc_swagger.yaml
swagger: '2.0'
info:
  description: >-
    Sugestão de documentação para ser utilizada como base para o desenvolvimento
    da API do projeto OPEN PMO da SEP-ES. 
  version: 1.0.0
  title: Open PMO Api
  termsOfService: 'http://www.totemti.com.br/'
  contact:
    email: atendimento@totemti.com.br
host: localhost
basePath: /openPmo
...
schemes:
  - https
  - http
paths:
...
  '/offices/{id}':
    get:
      tags:
        - officeController
      summary: Get office
      description: ''
      operationId: getOffice
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: id
          required: true
          type: integer
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/OfficeDto'
        '400':
          description: Bad Request
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error
    
    delete:
      tags:
        - officeController
      summary: Delete an office
      description: ''
      operationId: deleteOffice
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: id
          required: true
          type: integer
      responses:
        '200':
          description: Ok
        '400':
          description: Invalid input
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error
...
  '/workpacks/{id}':
    get:
      tags:
        - workpackController
      summary: Get workpack from a plan
      description: ''
      operationId: getWorkpack
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: id from workpack
          required: true
          type: integer
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/WorkpackDetailDto'
        '400':
          description: Bad Request
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error

    delete:
      tags:
        - workpackController
      summary: Delete a workpack from a plan
      description: ''
      operationId: deleteWorkpack
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: id from workpack
          required: true
          type: integer
      responses:
        '200':
          description: Ok
        '400':
          description: Invalid input
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error
...
  '/schedules/{id}':
    get:
      tags:
        - scheduleController
      summary: Get a schedule from a workpack
      description: ''
      operationId: getSchedule
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: id from schedule
          required: true
          type: integer
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/ScheduleDto'
        '400':
          description: Bad Request
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error

    delete:
      tags:
        - scheduleController
      summary: Delete a schedule from a workpack
      description: ''
      operationId: deleteSchedule
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: id from schedule
          required: true
          type: integer
      responses:
        '200':
          description: Ok
        '400':
          description: Invalid input
        '401':
          description: Not authenticate
        '500':
          description: Internall Server Error
...
securityDefinitions:
  api_key:
    type: apiKey
    name: api_key
    in: header
definitions:
...
  OfficeDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      fullName:
        type: string
...
  WorkpackDetailDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
      type:
        type: string
        description: Type of Workpack
        enum:
          - Portfolio
          - Program
          - Organizer
          - Deliverable
          - Project
          - Milestone
      workpackModel:
        $ref: '#/definitions/WorkpackModelDetailDto'
      properties:
        type: array
        items:
          $ref: '#/definitions/PropertyDto'
      children:
        type: array
        items:
          $ref: '#/definitions/WorkpackModelDto'
...
  PropertyDto:
    type: object
    discriminator: type
    required:
      - type
      - idPropertyModel
    properties:
      id:
        type: integer
        format: int64
      type:
        type: string
        description: Type of Workpack Model
        enum:
          - Integer
          - Text
          - Date
          - Toggle
          - UnitSelection
          - Selection
          - TextArea
          - Number
          - Currency
          - LocalitySelection
          - OrganizationSelection
      idPropertyModel:
        type: integer
        format: int64
...
  WorkpackModelDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
      type:
        type: string
        description: Type of Workpack Model
        enum:
          - PortfolioModel
          - ProgramModel
          - OrganizerModel
          - DeliverableModel
          - ProjectModel
          - MilestoneModel
      modelName:
        type: string
      fontIcon:
        type: string
...
  WorkpackModelDetailDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
      idPlanModel:
        type: integer
        format: int64
      idParent:
        type: integer
        format: int64
      type:
        type: string
        description: Type of Workpack Model
        enum:
          - PortfolioModel
          - ProgramModel
          - OrganizerModel
          - DeliverableModel
          - ProjectModel
          - MilestoneModel
      modelName:
        type: string
      modelNameInPlural:
        type: string
      fontIcon:
        type: string
      sortBy:
        type: string
      costSessionActive:
        type: boolean
      stakeholderSessionActive:
        type: boolean
      scheduleSessionActive:
        type: boolean
      childWorkpackModelSessionActive:
        type: boolean
      personRoles:
        type: array
        items:
          type: string
      organizationRoles:
        type: array
        items:
          type: string
      properties:
        type: array
        items:
          $ref: '#/definitions/PropertyModelDto'
...
  PropertyModelDto:
    type: object
    discriminator: type
    required:
      - type
    properties:
      id:
        type: integer
        format: int64
      type:
        type: string
        description: Type of Workpack Model
        enum:
          - IntegerModel
          - TextModel
          - DateModel
          - ToggleModel
          - UnitSelectionModel
          - SelectionModel
          - TextAreaModel
          - NumberModel
          - CurrencyModel
          - LocalitySelectionModel
          - OrganizationSelectionModel
      name:
        type: string
      label:
        type: string
      sortIndex:
        type: integer
      fullLine:
        type: boolean
      required:
        type: boolean
...
  ScheduleDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
      start:
        description: Date of pattern yyyy/MM/dd
        type: string
      finish:
        description: Date of pattern yyyy/MM/dd
        type: string
      plannedWork:
        type: number
      actualWork:
        type: number
      costs:
        type: array
        items:
          $ref: '#/definitions/StepCostParamDto'
...
  StepCostParamDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
      plannedCost:
        type: number
      actualCost:
        type: number
...